Split files for JS components between `helpers` and `annotations`, to mimic what was done for TS components
