--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/javascript/jsdoc-uxpin-annotations.ts
@@ -1,75 +1,0 @@
-import { CommentTags } from '../../CommentTags';
-import { getCommentTag } from '../../comments/getCommentTag';
-import { parseUsePortal } from '../../comments/parseUsePortal';
-import { ComponentNamespace } from './../../ComponentDefinition';
-import { getComponentNamespaceImportSlug } from '../getComponentNamespaceImportSlug';
-
-export function getComponentDocUrlFromJsDocTags(jsDocTags: string[]) {
-  return extractSingleWordFromJsDocTags(CommentTags.UXPIN_DOC_URL, jsDocTags);
-}
-
-export function getComponentNamespaceFromJsDocTags(
-  componentName: string,
-  jsDocTags: string[]
-): ComponentNamespace | undefined {
-  const namespaceName = extractSingleWordFromJsDocTags(CommentTags.UXPIN_NAMESPACE, jsDocTags);
-  if (!namespaceName) {
-    return;
-  }
-  return {
-    importSlug: getComponentNamespaceImportSlug(namespaceName, componentName),
-    name: namespaceName,
-  };
-}
-
-/**
- * Extract the value for `@uxpinuseportal` annotation from all JSDoc annotations
- *  can be either undefined (not found), true (found with no value) or a condition such as `props.mode === "inline"`
- */
-export function getComponentUsePortalFromJsDocTags(jsDocTags: string[]) {
-  const tagName = CommentTags.UXPIN_USE_PORTAL;
-  const description = getCommentTag(tagName, jsDocTags);
-  if (!description) {
-    return;
-  }
-  const inlineValue = extractMultipleWordFromJsDocTags(tagName, jsDocTags);
-  if (!inlineValue) {
-    return true;
-  }
-  return parseUsePortal(inlineValue);
-}
-
-export function extractSingleWordFromJsDocTags(tagName: CommentTags, jsDocTags: string[]): string | undefined {
-  const foundLineWithTag = getCommentTag(tagName, jsDocTags) || '';
-  return getCommentTagSingleWord(foundLineWithTag, tagName);
-}
-
-export function getCommentTagSingleWord(comment: string, tag: string): string | undefined {
-  if (!hasCommentTag(comment, tag)) {
-    return;
-  }
-  const pattern = new RegExp(`${escapeForRegexp(tag)}\\s+([a-z0-9\\.\\_\\-\\:\\/\\#]+)\\s*$`, 'im');
-  const matches: string[] | null = comment.match(pattern);
-  if (!matches) {
-    return;
-  }
-  return matches[1];
-}
-
-export function hasCommentTag(comment: string, tag: string): boolean {
-  return comment.includes(tag);
-}
-
-export function extractMultipleWordFromJsDocTags(tagName: CommentTags, jsDocTags: string[]): string | undefined {
-  const description = getCommentTag(tagName, jsDocTags);
-  if (!description) {
-    return;
-  }
-  const firstLine = description.split('\n')[0];
-  const inlineValue: string | undefined = firstLine.slice(tagName.length).trim();
-  return inlineValue;
-}
-
-function escapeForRegexp(string: string): string {
-  return string.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
-}
